chore: split withFocusOnFirstError to a new file
